Refactor main process initialization to reduce complexity
