+++ main/ui/menu.ts
@@ -1,0 +1,79 @@
+import { Menu, app, BrowserWindow } from 'electron';
+
+export function createMenu(getWindow: () => BrowserWindow | null) {
+  const template: Electron.MenuItemConstructorOptions[] = [
+    {
+      label: 'File',
+      submenu: [
+        {
+          label: 'Exit',
+          accelerator: process.platform === 'darwin' ? 'Cmd+Q' : 'Ctrl+Q',
+          click: () => {
+            app.quit();
+          },
+        },
+      ],
+    },
+    {
+      label: 'Import',
+      submenu: [
+        {
+          label: 'Add Game',
+          accelerator: 'Ctrl+N',
+          click: () => {
+            getWindow()?.webContents.send('menu:addGame');
+          },
+        },
+        {
+          type: 'separator',
+        },
+        {
+          label: 'Scan Folder for Games',
+          click: () => {
+            getWindow()?.webContents.send('menu:scanFolder');
+          },
+        },
+        {
+          label: 'Update Steam Library',
+          click: () => {
+            getWindow()?.webContents.send('menu:updateSteamLibrary');
+          },
+        },
+        {
+          type: 'separator',
+        },
+        {
+          label: 'Configure Steam...',
+          click: () => {
+            getWindow()?.webContents.send('menu:configureSteam');
+          },
+        },
+      ],
+    },
+    {
+      label: 'Edit',
+      submenu: [
+        { role: 'undo', label: 'Undo' },
+        { role: 'redo', label: 'Redo' },
+        { type: 'separator' },
+        { role: 'cut', label: 'Cut' },
+        { role: 'copy', label: 'Copy' },
+        { role: 'paste', label: 'Paste' },
+      ],
+    },
+    {
+      label: 'Help',
+      submenu: [
+        {
+          label: 'Check for Updates',
+          click: () => {
+            getWindow()?.webContents.send('menu:checkForUpdates');
+          },
+        },
+      ],
+    },
+  ];
+
+  const menu = Menu.buildFromTemplate(template);
+  Menu.setApplicationMenu(menu);
+}
